requirements.txt
glitch-this==1.0.0
Pillow==8.0.0
PyAutoGUI==0.9.52
